remove rules_from_lists from public API
