+++ benches/bench_matching.rs
@@ -6,7 +6,10 @@ use adblock::blocker::{Blocker, BlockerOptions};
 use adblock::engine::Engine;
 use adblock::request::Request;
 use adblock::url_parser::parse_url;
-use adblock::utils::rules_from_lists;
+
+#[path = "../tests/test_utils.rs"]
+mod test_utils;
+use test_utils::rules_from_lists;
 
 #[allow(non_snake_case)]
 #[derive(Serialize, Deserialize, Clone)]
