+++ benches/bench_rules.rs
@@ -2,7 +2,10 @@ use criterion::*;
 use once_cell::sync::Lazy;
 
 use adblock::blocker::{Blocker, BlockerOptions};
-use adblock::utils::rules_from_lists;
+
+#[path = "../tests/test_utils.rs"]
+mod test_utils;
+use test_utils::rules_from_lists;
 
 static DEFAULT_LISTS: Lazy<Vec<String>> = Lazy::new(|| {
     rules_from_lists(&[
